+++ tests/test_safeguard.py
@@ -1,25 +1,90 @@
-from chaoslib.exceptions import InterruptExecution
+from chaoslib.exceptions import InvalidActivity
 import pytest
 
-from chaosaddons.controls.safeguards import configure_control
+from chaosaddons.controls.safeguards import validate_control
 
 
 def test_fail_on_invalid_probes():
-    missing_type_probe = {"name": "hello"}
-    with pytest.raises(InterruptExecution) as x:
-        configure_control(probes=[missing_type_probe])
-        assert isinstance(x, InterruptExecution)
+    invalid_type_probe = {
+        "name": "my control",
+        "provider": {
+            "type": "python",
+            "module": "chaosaddons.controls.safeguards",
+            "arguments": [
+                {
+                    "name": "my probe",
+                    "type": "action",  ## should be a probe
+                    "provider": {
+                        "type": "python",
+                        "module": "os.path",
+                        "func": "exists"
+                    }
+                }
+            ]
+        }
+    }
+    with pytest.raises(InvalidActivity) as x:
+        validate_control(invalid_type_probe)
 
 
 def test_fail_on_invalid_probes_with_unknown_python_function():
     invalid_python_func_probe = {
-        "name": "hello",
-        "type": "probe",
+        "name": "my control",
         "provider": {
             "type": "python",
-            "module": "os",
-            "func": "whatever"
-        }}
-    with pytest.raises(InterruptExecution) as x:
-        configure_control(probes=[invalid_python_func_probe])
-        assert isinstance(x, InterruptExecution)
+            "module": "chaosaddons.controls.safeguards",
+            "arguments": [
+                {
+                    "name": "my probe",
+                    "type": "probe",
+                    "provider": {
+                        "type": "python",
+                        "module": "os.path",
+                        "func": "whatever"
+                    }
+                }
+            ]
+        }
+    }
+    with pytest.raises(InvalidActivity) as x:
+        validate_control(invalid_python_func_probe)
+
+
+def test_fail_on_missing_tolerance():
+    invalid_python_func_probe = {
+        "name": "my control",
+        "provider": {
+            "type": "python",
+            "module": "chaosaddons.controls.safeguards",
+            "arguments": [
+                {
+                    "name": "my probe",
+                    "type": "probe",
+                    "provider": {
+                        "type": "python",
+                        "module": "os.path",
+                        "func": "exists",
+                        "arguments": {
+                            "path": "/tmp"
+                        }
+                    }
+                }
+            ]
+        }
+    }
+    with pytest.raises(InvalidActivity) as x:
+        validate_control(invalid_python_func_probe)
+
+
+def test_fail_when_no_probes_were_given():
+    invalid_python_func_probe = {
+        "name": "my control",
+        "provider": {
+            "type": "python",
+            "module": "chaosaddons.controls.safeguards",
+            "arguments": [
+            ]
+        }
+    }
+    with pytest.raises(InvalidActivity) as x:
+        validate_control(invalid_python_func_probe)